requirements.txt
azure-storage-blob==12.25.1
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.10.6
pytest==8.3.5
streamlit==1.44.1
opencv-python-headless==4.10.0.84
